Add triton gemm test.
